+++ src/job.rs
@@ -234,11 +234,8 @@ mod test {
 
         let client = Client::from_service_account_key_file(sa_key).await;
 
-        // Delete the dataset if needed
-        let result = client.dataset().delete(project_id, dataset_id, true).await;
-        if let Ok(_) = result {
-            println!("Removed previous dataset '{}'", dataset_id);
-        }
+        client.table().delete_if_exists(project_id, dataset_id, table_id).await;
+        client.dataset().delete_if_exists(project_id, dataset_id, true).await;
 
         // Create dataset
         let created_dataset = client.dataset().create(Dataset::new(project_id, dataset_id)).await?;
@@ -345,10 +342,12 @@ mod test {
             },
         )?;
 
-        client
+        let result = client
             .tabledata()
             .insert_all(project_id, dataset_id, table_id, insert_request)
-            .await?;
+            .await;
+
+        assert!(result.is_ok(), "{:?}", result);
 
         // Query
         let mut rs = client