+++ src/model/category_count.rs
@@ -1,6 +1,6 @@
 //! Represents the count of a single category within the cluster.
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct CategoryCount {
     /// The name of category.