+++ src/model/data_format_options.rs
@@ -1,4 +1,4 @@
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct DataFormatOptions {
     /// Output timestamp as usec int64. Default is false.