+++ src/model/table_data_insert_all_response_insert_errors.rs
@@ -1,7 +1,8 @@
 use crate::model::error_proto::ErrorProto;
 use serde::{Deserialize, Serialize};
+use std::fmt::{Display, Formatter};
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct TableDataInsertAllResponseInsertErrors {
     /// Error information for the row indicated by the index property.
@@ -10,4 +11,10 @@ pub struct TableDataInsertAllResponseInsertErrors {
     /// The index of the row that error applies to.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub index: Option<i32>,
+}
+
+impl Display for TableDataInsertAllResponseInsertErrors {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.write_str(&format!("{:?}", self))
+    }
 }