+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.10.0] - 2021-10-09
+
+### Added
+
+- 2 features `native-tls` and `rust-tls` to respectively use OpenSSL or Rust TLS.
+- Add methods `is_empty`, `len`, `clear` to `TableDataInsertAllRequest`
+- Add method `add_rows` to `TableDataInsertAllRequest` (thanks @nixxholas)
+- Bump version for yup-oauth2 v6 (thanks @JamesHinshelwood)
+- Implement `Default` trait for most of the structures in the model sub-directory
+- Implement `Clone` trait for most of the structures in the model sub-directory
+- Implement `Display` trait for `ErrorProto` and `TableDataInsertAllResponseInsertErrors`
+- 
 ## [0.9.3] - 2021-08-31
 
 ### Fix